+++ types/adapter/repositories/GraphqlProjectRepository.d.ts
@@ -1,7 +1,7 @@
 import { BaseGitHubRepository } from './BaseGitHubRepository';
 import { ProjectRepository } from '../../domain/usecases/adapter-interfaces/ProjectRepository';
 import { Project } from '../../domain/entities/Project';
-export declare class GraphqlProjectRepository extends BaseGitHubRepository implements Pick<ProjectRepository, 'getProject' | 'findProjectIdByUrl'> {
+export declare class GraphqlProjectRepository extends BaseGitHubRepository implements Pick<ProjectRepository, 'getProject' | 'findProjectIdByUrl' | 'getByUrl'> {
     extractProjectFromUrl: (projectUrl: string) => {
         owner: string;
         projectNumber: number;
@@ -9,5 +9,6 @@ export declare class GraphqlProjectRepository extends BaseGitHubRepository imple
     fetchProjectId: (login: string, projectNumber: number) => Promise<string>;
     findProjectIdByUrl: (projectUrl: string) => Promise<Project["id"] | null>;
     getProject: (projectId: Project["id"]) => Promise<Project | null>;
+    getByUrl: (url: string) => Promise<Project>;
 }
 //# sourceMappingURL=GraphqlProjectRepository.d.ts.map