+++ types/domain/entities/ClaudeWindowUsage.d.ts
@@ -1,0 +1,6 @@
+export type ClaudeWindowUsage = {
+    hour: number;
+    utilizationPercentage: number;
+    resetsAt: Date;
+};
+//# sourceMappingURL=ClaudeWindowUsage.d.ts.map