+++ types/domain/entities/Comment.d.ts
@@ -1,0 +1,6 @@
+export type Comment = {
+    author: string;
+    content: string;
+    createdAt: Date;
+};
+//# sourceMappingURL=Comment.d.ts.map