+++ types/domain/entities/Project.d.ts
@@ -7,6 +7,7 @@ export type FieldOption = {
 };
 export type Project = {
     id: string;
+    url: string;
     databaseId: number;
     name: string;
     status: {