+++ types/domain/usecases/adapter-interfaces/ClaudeRepository.d.ts
@@ -1,0 +1,5 @@
+import { ClaudeWindowUsage } from '../../entities/ClaudeWindowUsage';
+export interface ClaudeRepository {
+    getUsage(): Promise<ClaudeWindowUsage[]>;
+}
+//# sourceMappingURL=ClaudeRepository.d.ts.map